app/requirements.txt
Django>=3.0.7
psycopg2-binary>=2.8.5
djangorestframework>=3.11.0
Pygments>=2.6.1
Markdown>=3.2.1
coreapi>=2.3.3
PyYAML>=5.3.1
django-elasticsearch-dsl-drf>=0.20.8
pyjwt>=1.7.1
python-decouple>=3.3